+++ src/nationalarchives/all+analytics.mjs
@@ -1,10 +1,9 @@
 import {
-  initAll,
-  Cookies,
   Accordion,
   Breadcrumbs,
   CodeBlock,
   CookieBanner,
+  Cookies,
   DateInputProgressive,
   ErrorSummary,
   FileInputDroppable,
@@ -13,10 +12,12 @@ import {
   GlobalHeader,
   Header,
   Picture,
+  Sidebar,
   SkipLink,
   Tabs,
-  TextInputPassword,
   TextAreaItemisedRows,
+  TextInputPassword,
+  initAll,
 } from "./all.mjs";
 import { EventTracker, GA4, helpers } from "./analytics.mjs";
 
@@ -35,6 +36,7 @@ export {
   GlobalHeader,
   Header,
   Picture,
+  Sidebar,
   SkipLink,
   Tabs,
   TextInputPassword,