+++ src/nationalarchives/components/search-field/analytics.js
@@ -10,7 +10,6 @@ export default [
         targetElement: ".tna-search-field__input",
         on: "blur",
         data: {
-          // eslint-disable-next-line no-unused-vars
           value: valueGetters.value,
           // eslint-disable-next-line no-unused-vars
           group: ($el, $scope, event) =>