+++ src/nationalarchives/templates/emails.stories.js
@@ -1,5 +1,6 @@
-import Template from "./layouts/email.njk?raw";
 import nunjucks from "nunjucks";
+
+import Template from "./layouts/email.njk?raw";
 
 nunjucks.configure(import.meta.env.PROD ? "" : "src");
 
@@ -9,13 +10,11 @@ export default {
     subject: { control: "text" },
     panel: { control: "object" },
     content: { control: "text" },
-    cta_buttons: { control: "object" },
+    ctaButtons: { control: "object" },
     signoff: { control: "text" },
     disableAdornableLogo: { control: "boolean" },
   },
-  render: (params) => {
-    return nunjucks.renderString(Template, { ...params });
-  },
+  render: (params) => nunjucks.renderString(Template, { ...params }),
 };
 
 export const Basic = {
@@ -40,6 +39,7 @@ export const Confirmation = {
       body: "<p>Your reference number is <strong>TNA1756716690J57</strong>.</p>",
     },
     content:
+      /* eslint-disable-next-line no-secrets/no-secrets */
       '<p>Thank you for contacting The National Archives with your application to film or record here.</p><p>You will receive a response from us about your application within 10 working days.</p><p>The details you have provided to us are listed below. If any of these are incorrect or if you have any questions regarding your enquiry, please email us at <a href="#">enquiry@nationalarchives.gov.uk</a> or call <a href="#">+44 (0) 20 8876 3444</a>, quoting your reference number.</p>',
     disableAdornableLogo: true,
   },
@@ -53,7 +53,7 @@ export const CallsToAction = {
     subject: "Respond to a request to access a record",
     content:
       '<p>You have received a request to access the following record:</p><dl class="tna-dl"><dt>Record IAID</dt><dd>C1234</dd><dt>Record name</dt><dd>Example record</dd></dl><p>Please review the request and either accept or reject it.</p>',
-    cta_buttons: [
+    ctaButtons: [
       {
         text: "Accept this request",
         href: "#",